requirements.txt
alembic
apscheduler
fastapi
feedparser
httpx
markdown
pydantic
pyyaml
python-dotenv
resend
sqlalchemy
uvicorn[standard]
aiosqlite
tzlocal
bleach
